Moved the api router to the 'routes' directory  and updated log messages
